+++ lib/bitstyles_phoenix/components/form.ex
@@ -56,6 +56,20 @@ defmodule BitstylesPhoenix.Form do
       ~s(<label for="user_name">Name</label><input id="user_name" maxlength="255" name="user[name]" type="text">)
   """)
 
+  story("Text field with error", """
+      iex> safe_to_string ui_input(@error_form, :name)
+      ~s(<label for="user_name">Name</label><input id="user_name" maxlength="255" name="user[name]" type="text"><span class="u-fg--warning" phx-feedback-for="user_name">is too short</span>)
+  """)
+
+  story("Text field with multiple errors", """
+      iex> safe_to_string ui_input(@error_form, :email)
+      ~s(<label for="user_email">Email</label><input id="user_email" maxlength="255" name="user[email]" type="text"><ul class=\"u-padding-l-left u-fg--warning\">
+        <li><span class=\"u-fg--warning\" phx-feedback-for=\"user_email\">is invalid</span></li>
+        <li><span class=\"u-fg--warning\" phx-feedback-for=\"user_email\">must end with @bitcrowd.net</span></li>
+      </ul>
+      )
+  """)
+
   story("Text field with hidden label", """
       iex> safe_to_string ui_input(@user_form, :name, hidden_label: true)
       ~s(<label class="u-sr-only" for="user_name">Name</label><input id="user_name" maxlength="255" name="user[name]" type="text">)
@@ -166,9 +180,9 @@ defmodule BitstylesPhoenix.Form do
 
     label = PhxForm.label(form, field, label_text, label_opts)
     input = PhxForm.select(form, field, options, input_opts)
-    error = Error.ui_error(form, field)
+    errors = Error.ui_errors(form, field)
 
-    ~E"<%= label %><%= input %><%= error %>"
+    ~E"<%= label %><%= input %><%= errors %>"
   end
 
   defp unwrapped_input(form, field, opts) do
@@ -179,9 +193,9 @@ defmodule BitstylesPhoenix.Form do
 
     label = PhxForm.label(form, field, label_text, label_opts)
     input = render_input(type, form, field, input_opts)
-    error = Error.ui_error(form, field)
+    errors = Error.ui_errors(form, field)
 
-    ~E"<%= label %><%= input %><%= error %>"
+    ~E"<%= label %><%= input %><%= errors %>"
   end
 
   defp wrapped_input(form, field, opts) do
@@ -189,10 +203,10 @@ defmodule BitstylesPhoenix.Form do
     label_text = get_label(opts, field)
     input_opts = get_input_opts(opts)
 
-    error = Error.ui_error(form, field)
+    errors = Error.ui_errors(form, field)
     input = render_input(type, form, field, input_opts)
 
-    ~E"<%= PhxForm.label do %><%= input %><%= label_text %><% end %><%= error %>"
+    ~E"<%= PhxForm.label do %><%= input %><%= label_text %><% end %><%= errors %>"
   end
 
   defp render_input(:radio_button, form, field, opts) do