+++ CHANGELOG.md
@@ -5,9 +5,15 @@
 ### Added
 * Better documentation for form helpers
 
+### Changes
+* Multiple errors on ones field are now rendered as a list 
+
 ### Breaking/Bugfix
 * Do not forward options from form helpers to each `label` anymore, but instead pass options through `label_opts`.
 * Drop `hidden` and `reset` input types for `ui_input`.
+
+### Deprecations
+* deprecated `ui_error/2` in favour of the new `ui_errors/2`
 
 ## v0.4.0